add and list show for clients,projects & invoices are completed.

+++ lib/screen/invoice_screen/entities/invoice_model.dart
@@ -6,7 +6,7 @@ class Invoice {
   String? invoiceName;
   String? invoiceNumber;
   double? invoiceAmount;
-  int? projectvalue;
+  double? projectvalue;
   String? dueDate;
   String? projectLink;
   String? refrenceID;
